(Fix) regex for github archive
Getting the right version without the correct prefix is bad. It makes it
impossible to get the correct commit from github with the purl. For now it
is better to get just the full tag/branch/.. as version.

Getting the version with the regex from [semver](https://semver.org/#is-there-a-suggested-regular-expression-regex-to-check-a-semver-string):
  `^(?P<major>0|[1-9]\d*)\.(?P<minor>0|[1-9]\d*)\.(?P<patch>0|[1-9]\d*)(?:-(?P<prerelease>(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\.(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\+(?P<buildmetadata>[0-9a-zA-Z-]+(?:\.[0-9a-zA-Z-]+)*))?$`
an treat the rest as prefix is not always possible because not everyone uses semver.

Signed-off-by: Jürgen Löhel <juergen@loehel.de>